Modified M1.kill_tumor_cell from calling TC.apoptosis directly, to now setting its death_rpob to 1.
Before, a NoneType error would be raised for the position argument in a line in hte apoptosis method of TC.
I Believe this was raised because the propgramme did two attempts of killing the agent. First M1 did apoptosis on it, then it did it itself.
That is no longer the case